agents/requirements.txt
strands-agents[gemini]>=1.0.0
strands-agents-tools>=0.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
rembg>=2.0.0
vtracer>=0.6.0
